+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 - CUSTOMER DELIVERY : editable deliveryCondition
 - CRM : Addition of 'freeText' boolean in LostReason and 'lostReasonStr' in lead and opportunity.
 - Sale Order: if config is activated, create stock move after modifications validation.
+- Improve Manuf order calculation rules (allow to select if we want take into account the WAP price or cost price for 
+in progress manuf order and for bill of material, and add a prorata method for in progress manuf order calculation of cost sheet)
 
 ## Bug Fixes
 - StockMoveLine: Fix wrong price values in declaration of exchanges.
@@ -28,6 +30,9 @@
 - MANUF ORDER : error when finishMoAutomaticEmail = true
 - Stock Move Invoicing: get price from stock move lines instead of from product.
 - STOCKMOVELINE : Fix NPE
+- Partner Stock Settings: fix exception when there were no settings for a company.
+- Manuf Order: fix produced product quantities computation.
+- Fix NPE on finishing manuf order.
 - SaleOrder : Fixed error for action attrs to hide inAti field.
 
 ## [5.1.3] - 2019-02-21